requirements.txt
mysql-connector-python==8.0.30
python-dotenv==0.21.0
matplotlib==3.5.3
numpy==1.23.3
pandas==1.5.0
SQLAlchemy==1.4.41
pymysql==1.0.2
cryptography==38.0.1